Created separate file for routing node
